create objecthandling package and add GetAndAddObject method
